Backend/requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
pytesseract==0.3.9
Pillow==9.1.0
python-multipart